portal/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
